requirements.txt
google-adk
openai
openinference-instrumentation-google-adk
litellm
langfuse